requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication & Security
bcrypt==4.1.2
PyJWT==2.8.0
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0

# HTTP Client
requests==2.31.0
httpx==0.26.0

# Geospatial
geopy==2.4.1

# Mapping & Visualization
folium==0.15.1
streamlit==1.30.0
streamlit-folium==0.16.0

# NGROK
pyngrok==7.0.5

# Scheduling
apscheduler==3.10.4

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3